Improved config support in SDK and fixed commands
