+++ packages/sdk/pureml/cli/helpers.py
@@ -1,38 +1,34 @@
-from pathlib import Path
-from pureml.cli.puremlconfig import PureMLConfigYML
-from pureml.schema.backend import BackendSchema
-from pureml.schema.paths import PathSchema
-
+import json
+import os
+from pureml.schema import PathSchema
 
 path_schema = PathSchema().get_instance()
-backend_schema = BackendSchema().get_instance()
-project_path = Path.cwd()
-if Path.exists(project_path / "puremlconfig.yml"):
-    puremlconfig = PureMLConfigYML(project_path / "puremlconfig.yml")
-else:
-    puremlconfig = None
 
+def save_auth(org_id: str = None, access_token: str = None, email: str = None):
+    token_path = path_schema.PATH_USER_TOKEN
 
-def get_backend_base_url(backend_url: str = None):
-    if backend_url is not None and backend_url != "":
-        # override backend url (command specific option)
-        return backend_url
-    if puremlconfig is not None:
-        # user configured backend url (self-hosted or custom pureml backend instance)
-        backend_url = puremlconfig.data["backend_url"] if "backend_url" in puremlconfig.data else backend_schema.BACKEND_URL
+    token_dir = os.path.dirname(token_path)
+    os.makedirs(token_dir, exist_ok=True)
+
+    # Read existing token
+    if os.path.exists(token_path):
+        with open(token_path, "r") as token_file:
+            token = json.load(token_file)
+
+        if org_id is not None:
+            token["org_id"] = org_id
+        if access_token is not None:
+            token["accessToken"] = access_token
+        if email is not None:
+            if "email" in token and token["email"] != email:
+                token["org_id"] = ""
+            token["email"] = email
     else:
-        # default backend url (production cloud backend)
-        backend_url = backend_schema.BACKEND_URL
-    return backend_url
+        token = {"org_id": org_id, "accessToken": access_token, "email": email}
+        if org_id is None:
+            token["org_id"] = ""
 
-def get_frontend_base_url(frontend_url: str = None):
-    if frontend_url is not None and frontend_url != "":
-        # override frontend url (command specific option)
-        return frontend_url
-    if puremlconfig is not None:
-        # user configured frontend url (self-hosted or custom pureml frontend instance)
-        frontend_url = puremlconfig.data["frontend_url"] if "frontend_url" in puremlconfig.data else backend_schema.FRONTEND_URL
-    else:
-        # default frontend url (production cloud frontend)
-        frontend_url = backend_schema.FRONTEND_URL
-    return frontend_url
+    token = json.dumps(token)
+
+    with open(token_path, "w") as token_file:
+        token_file.write(token)
