+++ packages/sdk/pureml/cli/orgs.py
@@ -1,11 +1,14 @@
+from typing import Dict, List
 import requests
 import typer
 from rich import print
 from rich.console import Console
 from rich.table import Table
 from urllib.parse import urljoin
+from pureml.cli.helpers import save_auth
 from pureml.schema import BackendSchema, PathSchema
 from pureml.components import get_token
+from pureml.schema.backend import get_backend_base_url
 
 
 
@@ -26,10 +29,10 @@ def callback():
     """
 
 @app.command()
-def show():
+def show(backend_url: str = typer.Option("", "--backend-url", "-b", help="Backend URL for self-hosted or custom pureml backend instance")):
     access_token = get_token()
     url_path = "org"
-    url = urljoin(backend_schema.BASE_URL, url_path)
+    url = urljoin(get_backend_base_url(backend_url), url_path)
 
     headers = {
         "accept": "application/json",
@@ -43,28 +46,32 @@ def show():
         console = Console()
         count = 0
         table = Table("Sr.No.", "Name", "Description")
+        org_data = {}
         for org in org_all:
             count += 1
             table.add_row(str(count), org["org"]["name"], org["org"]["description"])
+            org_data[count] = {"id": org["org"]["uuid"], "name": org["org"]["name"]}
 
         console.print(table)
         print()
-        return count
+        return count, org_data
     else:
         print("[bold red]Unable to get the list of organizations!")
-        return None
+        return None, None
 
 @app.command()
-def select():
+def select(backend_url: str = typer.Option("", "--backend-url", "-b", help="Backend URL for self-hosted or custom pureml backend instance")):
     print("[bold green] Select the Sr.No. of Organization")
-    count = show()
+    count, org_data = show(backend_url=backend_url)
     if count:
         sr_no = -1
         while int(sr_no) not in range(1, count + 1):
             sr_no: str = typer.prompt("Enter your Sr.No. of Organization (1 .... " + str(count) + ")")
             if int(sr_no) not in range(1, count + 1):
                 print("[bold red]Invalid Sr.No. of Organization!")
-        return sr_no
+        save_auth(org_id=org_data[int(sr_no)]["id"])
+        selected_org_name = org_data[int(sr_no)]["name"]
+        print(f"[green]Successfully linked to organization {selected_org_name}!")
     else:
         print("[bold red]Did not Select any organization!")
         return None
